Syncstat support in local prod & misc servlet packaging cleanup.
 - Move web.xml to debian.
 - Split out verification and identity to separate servlets.
 - Misc cleanup in servlet puppet code.

There are still a bunch of improvements that can be made in this area.
Specifically,

 - Move Java code for different servlets out of src/spsv into different
   Java libraries.
 - Remove legacy puppet support for servlets & more misc puppet code
   cleanup.

Deferring this work to a later date.

---
Testing Gaps
---

1. I have not tested this using the old prod infrastructure, since we don't
   have a good way to do that. I'm going to let old CI take care of this.
   The majority of the puppet code is shared with private deployment so it
   would have been tested when I deployed to local prod, but there are indeed
   some untested corner cases. Fingers crossed.

2. Another gap in my testing is openid stuff. I made sure the servlet was
   running but I haven't actually gone through any workflows here.

+++ src/libservlet/src/com/aerofs/servlets/lib/db/jedis/PooledJedisConnectionProvider.java
@@ -22,7 +22,7 @@ public class PooledJedisConnectionProvider
         return _jedisPool.getResource();
     }
 
-    public void init_(String host, short port)
+    public void init_(String host, int port)
     {
         // Cannot call init more than once.
         assert _jedisPool == null;
